Resolve the stream_ctx wait conflict by retaining token-aware wait support and keeping CUDA synchronization non-throwing inside host-task callbacks.

+++ cub/test/catch2_test_warp_scan_api.cu
@@ -40,9 +40,9 @@ __global__ void InclusiveWarpScanKernel(int* output)
   // Allocate WarpScan shared memory for 4 warps
   __shared__ typename warp_scan_t::TempStorage temp_storage[num_warps];
 
-  int warp_id       = threadIdx.x / 32;
+  int warp_id       = static_cast<int>(threadIdx.x / 32);
   int initial_value = 3;
-  int thread_data   = threadIdx.x % 32 + warp_id;
+  int thread_data   = static_cast<int>(threadIdx.x % 32) + warp_id;
 
   // warp #0 input: {0, 1, 2, 3, ..., 31}
   // warp #1 input: {1, 2, 3, 4, ..., 32}
@@ -73,7 +73,7 @@ C2H_TEST("Warp array-based inclusive scan works with initial value", "[scan][war
 
   for (int i = 0; i < num_warps; ++i)
   {
-    auto start = expected.begin() + i * 32;
+    auto start = expected.begin() + i * 32; // NOLINT(bugprone-misplaced-widening-cast)
     auto end   = start + 32;
 
     cuda::std::iota(start, end, i); // initialize host input for every warp
@@ -92,7 +92,7 @@ __global__ void InclusiveWarpScanKernelAggr(int* output, int* d_warp_aggregate)
   // Allocate WarpScan shared memory for 4 warps
   __shared__ typename warp_scan_t::TempStorage temp_storage[num_warps];
 
-  int warp_id       = threadIdx.x / 32;
+  int warp_id       = static_cast<int>(threadIdx.x / 32);
   int initial_value = 3; // for each warp
   int thread_data   = 1;
   int warp_aggregate;
@@ -131,7 +131,7 @@ C2H_TEST("Warp array-based inclusive scan aggregate works with initial value", "
 
   for (int i = 0; i < num_warps; ++i)
   {
-    auto start   = expected.begin() + i * 32;
+    auto start   = expected.begin() + i * 32; // NOLINT(bugprone-misplaced-widening-cast)
     auto end     = start + 32;
     int init_val = 3;
 
@@ -139,8 +139,8 @@ C2H_TEST("Warp array-based inclusive scan aggregate works with initial value", "
 
     cuda::std::inclusive_scan(start, end, start, sum_op{}, init_val);
 
-    expected_aggr.push_back(expected[i * 32 + 31] - init_val); // warp aggregate does not take
-                                                               // initial value into account
+    expected_aggr.push_back(expected[i * 32 + 31] - init_val); // NOLINT(bugprone-misplaced-widening-cast) warp
+                                                               // aggregate does not take initial value into account
   }
 
   REQUIRE(expected == d_out);
@@ -166,10 +166,10 @@ __global__ void InclusiveWarpScanPartialKernel(int* output)
   // Allocate WarpScan shared memory for 4 warps
   __shared__ typename warp_scan_t::TempStorage temp_storage[num_warps];
 
-  int warp_id = threadIdx.x / 32;
+  int warp_id = static_cast<int>(threadIdx.x / 32);
 
   int initial_value = 3;
-  int thread_data   = threadIdx.x % 32 + warp_id;
+  int thread_data   = static_cast<int>(threadIdx.x % 32) + warp_id;
   if (threadIdx.x % 2 == 1)
   {
     thread_data = -thread_data;
@@ -206,7 +206,7 @@ C2H_TEST("Warp array-based partial inclusive scan works with initial value", "[s
 
   for (int i = 0; i < num_warps; ++i)
   {
-    auto start      = expected.begin() + i * 32;
+    auto start      = expected.begin() + i * 32; // NOLINT(bugprone-misplaced-widening-cast)
     auto end        = start + 32;
     int valid_items = cuda::std::clamp(40 - i * 16, 0, 32);
 
@@ -244,8 +244,8 @@ __global__ void InclusiveWarpScanPartialKernelAggr(int* output, int* d_warp_aggr
   // Allocate WarpScan shared memory for 4 warps
   __shared__ typename warp_scan_t::TempStorage temp_storage[num_warps];
 
-  int warp_id       = threadIdx.x / 32;
-  int lane_id       = threadIdx.x % 32;
+  int warp_id       = static_cast<int>(threadIdx.x / 32);
+  int lane_id       = static_cast<int>(threadIdx.x % 32);
   int initial_value = 3; // for each warp
   int thread_data   = (lane_id < num_warps - 1 - warp_id) ? 0 : 1;
   int valid_items   = 40 - warp_id * 16;
@@ -285,7 +285,7 @@ C2H_TEST("Warp array-based partial inclusive scan aggregate works with initial v
 
   for (int i = 0; i < num_warps; ++i)
   {
-    auto start      = expected.begin() + i * 32;
+    auto start      = expected.begin() + i * 32; // NOLINT(bugprone-misplaced-widening-cast)
     auto end        = start + 32;
     int valid_items = cuda::std::clamp(40 - i * 16, 0, 32);
     int init_val    = 3;
